requirements.txt
keras==2.10.0
matplotlib==3.6.0
numpy==1.23.3
pandas==1.5.0
Pillow==9.2.0
pycocotools==2.0.5
pyforge==1.3.0
pytorch_lightning==1.1.4
scikit_learn==1.1.2
tensorboard==2.10.0
tensorflow==2.10.0
tensorflow_datasets==4.7.0
timm==0.6.7
torch==1.10.1
torchvision==0.11.2
tqdm==4.64.1
wandb==0.13.3